+++ packages/components/atoms/f-link/src/tenants/en-AU.js
@@ -1,4 +1,10 @@
+const messages = {
+    ariaDescription: {
+        newLocation: 'Opens in a new window/screen/tab',
+        externalSite: 'Opens an external site in a new window/screen/tab'
+    }
+};
+
 export default {
-    locale: 'en-AU',
-    text: 'I am a VLink Component (AU)'
+    messages
 };